Temporarily remove references to asymmetric signatures
As discussed in #25, the current recommendations around asymmetric
signatures are flawed, and the signature scheme doesn't support storing
the key in an HSM (or even just having one key). Both of these flaws
should be fixed before we add back asymmetric signature support.

I purposely left many references to asymmetric signatures throughout the
text because the spec does support it, it just doesn't specify it yet.
The hope is to add the improved set of recommendations either in the
current version of the spec before the release or in the next version.