lib/java: fix release CI (use the new Central Portal method) (#222)
This was needed for releasing the Java library.

The gist of it: OSSRH is now deprecated, and we are forced to use the
new Central Portal method. The new method is unfortunately undocumented,
not widely supported, etc. So had to change from gradle to Maven in
order to use the only upstream supported plugin.

Fixes #220